requirements.txt
# Django y DRF - Base del framework
Django==4.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-environ==0.11.2

# Bases de datos y caché
psycopg2-binary==2.9.9
django-redis==5.4.0

# Comunicación en tiempo real (WebSocket) - Opcional
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.0

# Monitoreo y métricas
django-prometheus

# HTTP Client y utilidades
requests==2.32.3

# Servidor WSGI/Static files
gunicorn==21.2.0
whitenoise==6.7.0

# Autenticación y seguridad
pyJWT==2.9.0

# Celery para tareas asíncronas
celery[redis]>=5.3.0
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0

# Async handling
asgiref==3.7.2 
asyncpg==0.30.0

# Timezone handling
pytz

# API Documentation - Opcional (descomentar si necesitas Swagger)
# drf-yasg==1.21.7

# Monitoring - Opcional (descomentar si necesitas Flower para Celery)
# flower==2.0.1

# AWS services - Opcional (descomentar si necesitas)
# boto3==1.35.1
# botocore==1.35.1

# Additional utilities - Opcional (descomentar según necesidades)
# uvicorn==0.30.6        # ASGI server alternative